Visualization modes for rendering.
There are two ways of setting this right now. One is to set it on
startup using an environment varible.

QSG_VISUALIZE=

"batches" - Visualize batchtes in the renderer. Merged batches are
drawn with solid color and Unmerged batches are drawn with a diagonal
line pattern. Few unique colors means good batching. Unmerged batches
are bad if they contain many individual nodes.

"clip" - Visualize clipping as red areas on top of the scene.

"overdraw" - Visualize all items in 3D to highlight overdraws. This mode
can also be used to detect geometry outside the viewport to some
extent. Opaque items are rendered with a green tint while translucent
items are rendered with a red tint. The bounding box for the viewport
is rendered in blue.  Opaque content is easier for the scenegraph to
process and it can also be faster to render on some hardware.

"changes" - Changes in the scenegraph are visualized with a flashing
overlay with a random color. Changes on a primitive is visualized with
a solid color while changes in an ancestor, such as a matrix or
opacity changes is visualized with a pattern.

The second way to set the visualization mode is to set it at runtime
through QString QQuickWindowPrivate::customRenderMode. This "API" is
string based so it is not tied to the batch renderer and in theory can
support other custom renderers.

The visualized elements do not respect clipping and rendering order
is arbitrary.

Change-Id: I31efbe53fc905145bf48080ede3e36945cb60dcf
Reviewed-by: Michael Brasser <michael.brasser@live.com>

Conflicts:
	src/quick/items/qquickwindow.cpp
	src/quick/scenegraph/coreapi/qsgbatchrenderer.cpp
	src/quick/scenegraph/coreapi/qsgbatchrenderer_p.h
	src/quick/scenegraph/coreapi/qsgrenderer_p.h