QQuickItemParticle: delete delegates once they are completed.
This was one gigantic memory leak, if anyone tried to use this functionality, as
evidenced by the bug. The comment notes that deletion was only to be done if the
items were created by the ItemParticle, but I don't think this makes sense:
take()'s documentation doesn't imply that the ownership of the item stays with
the creator, and nor does this seem to be a particularly sensible API decision
to make, so I'm not going to implement that.

Change-Id: I307d77b70c2b4f7b34565865f8efcfe64817be2f
Task-number: QTBUG-37486
Reviewed-by: Gunnar Sletta <gunnar@sletta.org>