Introduce QSG_TRANSIENT_IMAGES.
This feature enables the same feature that we already have in the
'nonpreservedtexture' in the scene graph playground.

This change also cleans up some leftover logic from the early Qt 5.2
days when we had to manually extract the image from the default
texture factory rather than going through
QQuickWindow::createTexture(). This cleanup is required for the logic
to work.

Change-Id: I7cdf15d981ba8af94f8d51781fc01e2b834b77c6
Reviewed-by: Robin Burchell <robin.burchell@viroteck.net>