[qtdeclarative] Allow configuration to scale with dpi. Contributes to JB#43710
On runtime the active dpi is compared to "General/BaseDPI" on
QtQuick2.conf file to multiply configuration items value marked as
scalable. That is done by having configuration item name appended with
"Scaled", e.g.
[QuickFlickable]
FlickThresholdScaled=15

On MouseArea the old, not anymore used, configuration for press and
hold timeout was removed.

This commit can be squashed with earlier configuration support commit
when rebased to newer Qt releasese:
"Allow QtQuick2 parameters to be configured".