[qtdeclarative] Fix crash due to an interaction between an item layer and anchors. Fixes JB#44681
Evaluating anchors in QQuickItem::componentComplete() will change size
and position bindings and the right code reacting to those changes could
create a layer after QQuickItemPrivate::completeCreate is set to true
which means QQuickItemLayer::classBegin() would not be called, but
QQuickItemLayer::componentComplete() would be meaning
QQuickItemLayer::activate() would be called twice and it would be added
as an item change listener twice. If the layer was then ever deactivated
it would only be removed as a listener once and changes to the parent
size would update the layers size when it wasn't expecting it.