[qtdeclarative] Dump types reachable through internal dependencies. Contributes to JB#46800
Do not exclude types that are reachable through dependencies, but only
become registered (exported) after the dumped module is imported.  This
happens e.g. when a private submodule exists as a dependency and it
extends types from the public module. The extended types become
reachable when the private submodule is loaded as a dependency, but only
become registered (exported) after the public module is loaded.